+++ app/src/main/java/com/app/medoxnetwork/model/ModelWithdrawBalance.kt
@@ -5,6 +5,6 @@ data class ModelWithdrawBalance(
     val status: Int?
 ) {
     data class Result(
-        val balance: Int?
+        val balance: Double?
     )
 }
